services/db-import/requirements.txt
pypika==0.48.8
requests==2.26.0
python-dotenv==0.19.2